fixed a very subtle power drain bug that only showed up on some devices in sleep, and cleaned up some other missed details on this same subject.
main attraction:  GIO0 was changed to an input to accomodate the fact
that it was attached to a pulldown, meant for a button on the external
connector.

when an output, some boards burned ~180 ua sleeping, the others 630
ua.  as an input, we changed this reading on an example of the latter
to 180.

thanks to catalin caranfil at realtime oem for finding the problem and
pushing for a solution for the masses!

other changes:

- made onewire (data) input
- made the spi pins all gpio in appropriate directions for their
connections
- made bt_pio input, as appropriate for its connection