spi: fixing bug in writing delays and clocks of registers
For some reason, when one writes some parts of the chip select register, then other parts of the register get erased. Might be a bug. This fix takes care of this problem by first reading the register, changing the parts we need, and then writing it back into the register.