tc: things compile now, but doens't work just yet
- I added a test application in test/sam3u/timer/Alarm32khz to test the alarm
  component. It compiles, and does something, but it is not correct. Two issues:

1. For some reason, m_upper in TransformCounterC.nc:56 does not get
initialized to 0. Thus, we get an extremely high number to start off with.
Why? shouldn't that be in the BSS and thus get automatically zeroed?

2. If I force m_upper to 0 (edit code, but not checked in), then the alarm
starts at zero and it even fires. But, the fire is not at the right alarm. It
is at 2x of what I set the alarm to... no idea why!