tc: adding TMicro alarm
Added an alarm that tries as best as it can to get a TMicro alarm. However,
because of the 12MHz clock and PLL settings, this can not always achieve that
frequency. Therefore, there is a function in the HplSam3uTCChannel interface
to get the exact frequency of a timer in kHz.

See apps/tests/sam3u/timer/AlarmTMicro for an example on how to use it.