timer: Fixed misunderstanding of Capture unit
Each channel of the TC doesn't have two indipendent capture units. The two
capture registers are linked to each other and interchangably capture the
values. This is now correctly handled.

The disadvantage: We only have 3 capture units on the SAM3U
The advantage: we can capture events very fast without overflows. This will be
handy for the radio.